Add support for Casual vs. Competitive play.
This is in response to bug #79867 where the reporter wanted an easier
skill level.  According to a comment in the bug, this was fixed in KDE
3.2.2, by introducing a feature where the computer misses good moves
once in a while in a random fashion.

Unfortunately this also annoyed those who wanted KReversi to play as
strong as possible (me, among other people) since there was no way to
turn off this feature.

This patch gives the best of both worlds: Those who want to play
casually with a weak opponent turns on "casual play" in the settings
and those that want a tough SOB for an opponent turns on "Competitive
play".

Only games played in competitive mode will be entered into the
highscore list from now on.

This concept should also be applicable to other board games in
kdegames (kenolaba, kwin4, etc).

CCMAIL: 79867@bugs.kde.org

svn path=/trunk/kdegames/kreversi/; revision=347049