Implement wish 82519: Label the board with A-H, 1-8
This is the start of a modernization of the GUI for KReversi.

Unfortunately, I am not a great graphics designer, so anybody who
knows a nicer font is free to change it.  The code itself should be
flexible enough to adapt to any font.  Don't mess with the fontsize,
though.

FEATURE: 82519
GUI:

svn path=/trunk/kdegames/kreversi/; revision=401824