Start on wish 82900: Show possible moves in the current position
This commit implements the main parts of wish 82900.  The thing that
remains until I can close the bug is that there should be a toggle
action for this and also a new icon.

I am also aware that this feature doesn't work together with the "show
hint" function.

CCBUGS: 82900
GUI:

svn path=/trunk/kdegames/kreversi/; revision=402003