Bring KExtHighscore classes from libkdegames
KExtHighscore is being removed from libkdegames since they are considered deprecated.

KReversi its the only user so we're bringing the code here. This is an unfinished port that lived in libkdegames (see there's a few TODO). Would probably need some clearing also like removing the world-wide highscore support that we don't use and adds a considerable amount of code (i have a patch for that too)

Compiles, haven't done much more since it's the frameworks branch and thus work in progress/unreleased

REVIEW: 124393